making it indempotent

+++ recipes/logrotate.rb
@@ -30,12 +30,17 @@ Capistrano::Configuration.instance(true).load do
 
   desc "Move log directory to /mnt"
   task :move_log_dir_to_mnt do
-    sudo "mkdir -p /mnt/application_logs/#{application}"
-    sudo "chown -R #{user}:#{user} /mnt/application_logs/#{application}"
-    sudo "mv /var/www/apps/#{application}/shared/log/* /mnt/application_logs/#{application}"
-    sudo "rm -fr /var/www/apps/#{application}/shared/log"
-    sudo "ln -s /mnt/application_logs/#{application} /var/www/apps/#{application}/shared/log"
-    # Now need to deploy the app.
+    if Dir.exists?("/mnt/application_logs/#{application}")
+      # do nothing
+      puts "Doing nothing as the application log directory already exists in mount."
+    else
+      sudo "mkdir -p /mnt/application_logs/#{application}"
+      sudo "chown -R #{user}:#{user} /mnt/application_logs/#{application}"
+      sudo "mv /var/www/apps/#{application}/shared/log/* /mnt/application_logs/#{application}"
+      sudo "rm -fr /var/www/apps/#{application}/shared/log"
+      sudo "ln -s /mnt/application_logs/#{application} /var/www/apps/#{application}/shared/log"
+      # Now need to deploy the app.
+    end
   end
 
   end
